moved logic to new library
