+++ app/messenger.h
@@ -34,7 +34,7 @@ void MSG_EnableRX(const bool enable);
 void MSG_StorePacket(const uint16_t interrupt_bits);
 void MSG_Init();
 void MSG_ProcessKeys(KEY_Code_t Key, bool bKeyPressed, bool bKeyHeld);
-void MSG_Send(const char txMessage[TX_MSG_LENGTH], bool bServiceMessage);
+void MSG_Send(char txMessage[TX_MSG_LENGTH], bool bServiceMessage);
 
 #endif
